Initial chattings integration (#272)
By the wonderful help given by @0916dhkim in #262, we are now able to integrate chat features in our app. There are few typing issues beforehand and few resolvers are corrected which didn't return the required variables.

`messages` queries are newly added and this will be used to handle messages in [Message] screen.