Integrating messages query (#273)
Add [lastMessageId] to [Channel]. This will make client able to fetch channels in recent order which should be applied to [Channel] list items. [updatedAt] field will be refreshed when [lastMessageId] is updated.

Add filters to messages query when [searchText] is not null. If null is provided, the result is not what we wanted.

* Applied relay pagination to messages on client
* Impl adding newEdge when sending message