Fix profile status message overflow & subscription (#457)
There have been a issue #455 which profile status message shown in both `MainFriend` and `ProfileModal`. And I think it might be better if my friend updates his/her status message in `ProfileUpdate`, the status message which I'm seeing also should be updated in real-time.

Added and changed the codebase accordingly.