Release 1.9.0 and packages migration (#465)
Migrate all legacy packages and bump version to 1.9.0.

For the server-side, we'll preserve `Prisma` version to 2.13.+ and the apollo server to version 2 since there are breaking changes that will be resolved in the future. For `Prisma`, since we are still using old [nexus-plugin-prisma](https://github.com/graphql-nexus/nexus-plugin-prisma), we will hold it until new [nexus-prisma](https://github.com/prisma/nexus-prisma) release.